requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
fredapi>=0.5.1
newsapi-python>=0.2.7
transformers>=4.38.0
torch>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
